pyproject.toml
[project]
name = "multimodal-rag-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "fastapi>=0.121.2",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "openai>=2.8.1",
    "pdfplumber>=0.11.8",
    "pillow>=12.0.0",
    "prometheus-client>=0.23.1",
    "pypdf>=6.3.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "supabase>=2.0.0",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.41.0",
    "uvicorn>=0.38.0",
]

[project.scripts]
ingest-worker = "backend.worker:main"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "pre-commit>=4.3.0",
    # "pyright~=1.1.405",
    "pytest>=8.4.2",
    "ruff>=0.13.3",
]

[tool.ruff]
exclude = [
    ".git",
    ".ipynb_checkpoints",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "venv",
    "__pycache__",
]
fix = true
indent-width = 4
line-length = 79

[tool.ruff.lint]
select = [
    # "ANN",
    "B",
    "D",
    "E",
    "F",
    "I",
    "T",
]
fixable = ["ALL"]
# TODO: remove this
ignore = [
    "D100",
    "D104",
    "E501",
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true

# [tool.pyright]
# include = [
#     "multimodal-rag-gpt",
# ]
